requirements.txt
#pip install -r requirements.txt
streamlit==1.22.0
streamlit-chat==1.2.0
python-dotenv==1.0.0
openai==0.27.2
tenacity==8.2.2
pyodbc==4.0.35
pyodbc-pool==1.2.0
langchain==0.0.146
